// apps/web/src/api/model/billingItem.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Allawee Infrastructure Backend APIs
 * The Allawee Infrastructure APIs description
 * OpenAPI spec version: 0.0.1
 */
import type { BillingItemBillable } from './billingItemBillable';
import type { BillingItemUnit } from './billingItemUnit';

export interface BillingItem {
  /** Billable key */
  billable: BillingItemBillable;
  /** Product ID */
  product: string;
  /** Product reference type */
  productRef: string;
  /** Quantity */
  quantity: number;
  /** Billing unit */
  unit: BillingItemUnit;
  /** Unit price */
  unitPrice: number;
  /** Billable price */
  billablePrice: number;
}